Agregada la busqueda de palabras simples

+++ MoogleEngine/Occurrences.cs
@@ -3,7 +3,7 @@ namespace MoogleEngine;
 // Representa las ocurrencias de alguna palabra en un documento especifico
 public class Occurrences {
     
-    double tfidf;
+    float tfidf;
     List<int> startPos = new List<int> (); // Posiciones iniciales de cada ocurrencia
     
     public Occurrences(int id) {
@@ -15,12 +15,12 @@ public class Occurrences {
     }
 
     public int Id { get; private set; } // ID del documento analizado
-    public double Relevance { // La relevancia de la palabra en este documento
+    public float Relevance { // La relevancia de la palabra en este documento
         get {
             return tfidf;
         }
         set {
-            tfidf = (value == 0 ? 0.000001 : value);
+            tfidf = (value == 0 ? 0.000001f : value);
         }
     }
